add jack package, separate session logic
